Added facades, dtos,usecases, updated repository

+++ src/Services/BookingService/BookingService.Domain/Entities/SeatType.cs
@@ -5,5 +5,5 @@ namespace BookingService.Domain.Entities;
 public class SeatType:BaseEntity
 {
     public string Name { get; set; }
-    public double PriceMultiplier { get; set; }
+    public decimal PriceMultiplier { get; set; }
 }
